+++ Demo/Sources/Demos/Demo Screens/ReorderingViewController.swift
@@ -25,6 +25,7 @@ final class ReorderingViewController : UIViewController
         self.list.setContent { list in
             
             list.appearance = .demoAppearance
+            list.layout = .demoLayout
             
             list += Section(identifier: "first") { section in
                 section.header = HeaderFooter(DemoHeader(title: "First Section"))