+++ Demo/Sources/Demos/Demo Screens/SpacingCustomizationViewController.swift
@@ -23,14 +23,14 @@ final class SpacingCustomizationViewController : UIViewController
         self.view = self.listView
 
         self.listView.setContent { list in
-
-            list.appearance.list.layout.interSectionSpacingWithNoFooter = 16
-
-            list.appearance.list.layout.set {
-                $0.padding = UIEdgeInsets(top: 20.0, left: 20.0, bottom: 20.0, right: 20.0)
-                $0.itemSpacing = 20.0
-                $0.interSectionSpacingWithFooter = 20.0
-                $0.interSectionSpacingWithNoFooter = 20.0
+            
+            list.layout = .list {
+                $0.layout.set {
+                    $0.padding = UIEdgeInsets(top: 20.0, left: 20.0, bottom: 20.0, right: 20.0)
+                    $0.itemSpacing = 20.0
+                    $0.interSectionSpacingWithFooter = 20.0
+                    $0.interSectionSpacingWithNoFooter = 20.0
+                }
             }
 
             list += Section(identifier: "default") { section in