+++ Demo/Sources/Demos/Demo Screens/WidthCustomizationViewController.swift
@@ -24,11 +24,13 @@ final class WidthCustomizationViewController : UIViewController
         
         self.listView.setContent { list in
             
-            list.appearance.list.layout.set {
-                $0.padding = UIEdgeInsets(top: 20.0, left: 20.0, bottom: 20.0, right: 20.0)
-                $0.itemSpacing = 20.0
-                $0.interSectionSpacingWithFooter = 20.0
-                $0.interSectionSpacingWithNoFooter = 20.0
+            list.layout = .list {
+                $0.layout.set {
+                    $0.padding = UIEdgeInsets(top: 20.0, left: 20.0, bottom: 20.0, right: 20.0)
+                    $0.itemSpacing = 20.0
+                    $0.interSectionSpacingWithFooter = 20.0
+                    $0.interSectionSpacingWithNoFooter = 20.0
+                }
             }
             
             list += Section(identifier: "default") { section in